+++ demos/employees/src/main/java/module-info.java
@@ -30,7 +30,7 @@ module is.codion.demos.employees {
 	requires is.codion.swing.framework.ui;
 	requires is.codion.framework.server;
 	requires is.codion.plugin.jasperreports;
-	requires is.codion.plugin.intellij.themes;
+	requires is.codion.plugin.flatlaf.intellij.themes;
 
 	exports is.codion.demos.employees.domain
 					to is.codion.framework.domain, is.codion.framework.db.local;